양성일 : 맵 클린아키텍처로 변경

+++ Assets/UI_MapSetup.cs
@@ -12,7 +12,7 @@ public class UI_MapSetup : MonoBehaviour
         
         Hashtable roomProperties = new Hashtable()
         {
-            {ERoomProperties.MapSelected.ToString(), SelectedMap}
+            {ERoomProperties.MapSelected.ToString(), (int)SelectedMap}
         };
         
         PhotonNetwork.CurrentRoom.SetCustomProperties(roomProperties);
